experience - findById() and save()

+++ src/main/java/com/travelapp/serviceImpl/TripServiceImpl.java
@@ -1,7 +1,9 @@
 package com.travelapp.serviceImpl;
 
 import com.travelapp.entity.Trip;
+import com.travelapp.entity.User;
 import com.travelapp.exception.BadRequestException;
+import com.travelapp.exception.ForbiddenException;
 import com.travelapp.exception.NotFoundException;
 import com.travelapp.mapper.TripMapper;
 import com.travelapp.record.trip.CreateTripRecord;
@@ -41,29 +43,44 @@ public class TripServiceImpl implements TripService {
 
     @Override
     public TripResponseRecord findById(Long id) {
-        // later on add check if current user should be able to access this object
+        // later on get user from context
+        User currentUser = userRepository.findByUsername("ivana");
         Trip trip = tripRepository.findById(id).orElseThrow(() -> new NotFoundException("TripService findById() :: Trip not found with id " + id));
+        if(!trip.getUser().equals(currentUser)) {
+            throw new ForbiddenException("TripService findById() :: User " + currentUser.getUsername() + " does not have access to this trip");
+        }
         return tripMapper.toTripResponseRecord(trip);
     }
 
     @Override
     public TripResponseRecord save(CreateTripRecord createTripRecord) {
+        // later on get user from context
+        User currentUser = userRepository.findByUsername("ivana");
         Trip trip = tripMapper.toTrip(createTripRecord);
-        trip.setUser(userRepository.findById(createTripRecord.userId()).orElseThrow(() -> new BadRequestException("TripService save() :: Trip cannot be saved for there is no user with id " + createTripRecord.userId())));
+        trip.setUser(currentUser);
         return tripMapper.toTripResponseRecord(tripRepository.save(trip));
     }
 
     @Override
     public TripResponseRecord update(UpdateTripRecord updateTripRecord) {
-        // later on add check if current user should be able to update this object
+        // later on get user from context
+        User currentUser = userRepository.findByUsername("ivana");
         Trip trip = tripRepository.findById(updateTripRecord.id()).orElseThrow(() -> new NotFoundException("TripService update() :: Trip cannot be updated for there is no trip found with id " + updateTripRecord.id()));
+        if(!trip.getUser().equals(currentUser)) {
+            throw new ForbiddenException("TripService update() :: User " + currentUser.getUsername() + " does not have access to this trip");
+        }
         tripMapper.updateTripFromRecord(updateTripRecord, trip);
         return tripMapper.toTripResponseRecord(tripRepository.save(trip));
     }
 
     @Override
     public void deleteById(Long id) {
-        // later on add check if current user should be able to delete this object
+        // later on get user from context
+        User currentUser = userRepository.findByUsername("ivana");
+        Trip trip = tripRepository.findById(id).orElseThrow(() -> new NotFoundException("TripService delete() :: Trip cannot be deleted for there is no trip found with id " + id));
+        if(!trip.getUser().equals(currentUser)) {
+            throw new ForbiddenException("TripService delete() :: User " + currentUser.getUsername() + " does not have access to this trip");
+        }
         tripRepository.deleteById(id);
     }
 }
